# ref: SpanId creation