--- a/src/Sentry/Internal/RandomValuesFactory.cs
+++ b/src/Sentry/Internal/RandomValuesFactory.cs
@@ -6,6 +6,7 @@ internal abstract class RandomValuesFactory
     public abstract int NextInt(int minValue, int maxValue);
     public abstract double NextDouble();
     public abstract void NextBytes(byte[] bytes);
+    public abstract void NextBytes(Span<byte> bytes);
 
     public bool NextBool(double rate) => rate switch
     {