--- a/src/Sentry/Internal/RandomValuesFactory.cs
+++ b/src/Sentry/Internal/RandomValuesFactory.cs
@@ -6,7 +6,10 @@ internal abstract class RandomValuesFactory
     public abstract int NextInt(int minValue, int maxValue);
     public abstract double NextDouble();
     public abstract void NextBytes(byte[] bytes);
+
+#if NET6_0_OR_GREATER
     public abstract void NextBytes(Span<byte> bytes);
+#endif
 
     public bool NextBool(double rate) => rate switch
     {