--- a/src/Sentry/Internal/SynchronizedRandomValuesFactory.cs
+++ b/src/Sentry/Internal/SynchronizedRandomValuesFactory.cs
@@ -7,6 +7,7 @@ internal class SynchronizedRandomValuesFactory : RandomValuesFactory
         public override int NextInt(int minValue, int maxValue) => Random.Shared.Next(minValue, maxValue);
         public override double NextDouble() => Random.Shared.NextDouble();
         public override void NextBytes(byte[] bytes) => Random.Shared.NextBytes(bytes);
+        public override void NextBytes(Span<byte> bytes) => Random.Shared.NextBytes(bytes);
 #else
     private static readonly AsyncLocal<Random> LocalRandom = new();
     private static Random Random => LocalRandom.Value ??= new Random();
@@ -15,5 +16,10 @@ internal class SynchronizedRandomValuesFactory : RandomValuesFactory
     public override int NextInt(int minValue, int maxValue) => Random.Next(minValue, maxValue);
     public override double NextDouble() => Random.NextDouble();
     public override void NextBytes(byte[] bytes) => Random.NextBytes(bytes);
+    public override void NextBytes(Span<byte> bytes) => Random.NextBytes(bytes
+#if NETSTANDARD2_0 || NET461
+            .ToArray()
+#endif
+    );
 #endif
 }