--- a/src/Sentry/Internal/SynchronizedRandomValuesFactory.cs
+++ b/src/Sentry/Internal/SynchronizedRandomValuesFactory.cs
@@ -16,10 +16,5 @@ internal class SynchronizedRandomValuesFactory : RandomValuesFactory
     public override int NextInt(int minValue, int maxValue) => Random.Next(minValue, maxValue);
     public override double NextDouble() => Random.NextDouble();
     public override void NextBytes(byte[] bytes) => Random.NextBytes(bytes);
-    public override void NextBytes(Span<byte> bytes) => Random.NextBytes(bytes
-#if NETSTANDARD2_0 || NET461
-            .ToArray()
-#endif
-    );
 #endif
 }