--- a/src/Sentry/SpanId.cs
+++ b/src/Sentry/SpanId.cs
@@ -49,10 +49,10 @@ public readonly struct SpanId : IEquatable<SpanId>, IJsonSerializable
     /// </summary>
     public static SpanId Create()
     {
-#if NET6_0_OR_GREATER
-        Span<byte> buf = stackalloc byte[8];
-#else
+#if NETSTANDARD2_0 || NET461
         byte[] buf = new byte[8];
+#else
+        Span<byte> buf = stackalloc byte[8];
 #endif
 
         Random.NextBytes(buf);
@@ -73,12 +73,13 @@ public readonly struct SpanId : IEquatable<SpanId>, IJsonSerializable
         Span<byte> convertedBytes = stackalloc byte[sizeof(long)];
         Unsafe.As<byte, long>(ref convertedBytes[0]) = _value;
 
+        // Going backwards through the array to preserve the order of the output hex string (i.e. `4e76` -> `76e4`)
         Span<char> output = stackalloc char[16];
-        for (var i = 0; i < convertedBytes.Length; i++)
+        for (var i = convertedBytes.Length - 1; i >= 0; i--)
         {
             var value = convertedBytes[i];
-            output[i * 2] = HexChars[value >> 4];
-            output[i * 2 + 1] = HexChars[value & 0xF];
+            output[(convertedBytes.Length - 1 - i) * 2] = HexChars[value >> 4];
+            output[(convertedBytes.Length - 1 - i) * 2 + 1] = HexChars[value & 0xF];
         }
 
         writer.WriteStringValue(output);