--- a/test/Sentry.Testing/IsolatedRandomValuesFactory.cs
+++ b/test/Sentry.Testing/IsolatedRandomValuesFactory.cs
@@ -11,7 +11,8 @@ internal class IsolatedRandomValuesFactory : RandomValuesFactory
     public override double NextDouble() => _random.NextDouble();
 
     public override void NextBytes(byte[] bytes) => _random.NextBytes(bytes);
-#if NET6_0_OR_GREATER
+
+#if !(NETSTANDARD2_0 || NET48)
     public override void NextBytes(Span<byte> bytes) => _random.NextBytes(bytes);
 #endif