In an effort to improve the memory footprint: There is no technical reason to keep the Id value as `string`.
The spanId value of `0` will be treated as `empty`.

```
BenchmarkDotNet=v0.13.5, OS=macOS Ventura 13.4 (22F66) [Darwin 22.5.0]
Apple M1 Pro, 1 CPU, 10 logical and 10 physical cores

OLD:
|              Method |     Mean |    Error |  StdDev |   Gen0 | Allocated |
|-------------------- |---------:|---------:|--------:|-------:|----------:|
| 'Creates a Span ID' | 722.1 ns | 63.51 ns | 3.48 ns | 0.0687 |     144 B |

NEW:
|              Method |     Mean |     Error |    StdDev | Allocated |
|-------------------- |---------:|----------:|----------:|----------:|
| 'Creates a Span ID' | 7.924 ns | 0.3097 ns | 0.0170 ns |         - |
```

#skip-changelog